fix: syntax error

+++ src/js/order_ok.js
@@ -81,6 +81,8 @@ const orderOk = new OrderOk(deliveryHistory)
 
 orderOk.render();
 
-homeBtn.addEventListener('click', ()=> {
-    localStorage.clear();
-})
+if(localStorage.getItem('uuid')){
+    homeBtn.addEventListener("click", () => {
+        localStorage.clear();
+    });
+}
